Force disconnect connections to primary db only
Database connections are forcibly disconnected after a default of one
hour to prevent an imbalance of connections from the application to the
pool of PgBouncers.

However this only applies for connections targeting the primary, which
sits behind multiple PgBouncers. Change the disconnection logic to only
forcibly disconnect when the connection is to the primary database.