Fix retry-lock problem on pipeline cancel
Having the "retries" parameter nil causes an error on
Gitlab::OptimisticLocking when comparing retry_attempts < max_retries.
So we need to change its default value. We can make it `0`,
but I think it's better to have `1` to have a "secure" cancel mechanism.

The retry lock handles all jobs in the current implementation,
this MR also changes the lock mechanism to an individual job.

Changelog: fixed