Upgrade gitaly dependency to use proper go modules versioning
Gitaly project now properly respects module release flow
and includes a module suffix in the package name. It requires
to re-write all non-suffixed imports with suffixed of a specific
version of tha module. With proper module versioning we don't
need to use a 'replace' directive to point to specific commit
and can use semantic versioning for the gitaly dependency.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/3177