Refactor db selection in SidekiqServerMiddleware
This was all happening in a side-effecting getter
which made it difficult to reason about. It also
meant that we sometimes updated the job hash with
the selected database and consistency setting and
other times we did not.

I split up these two concerns so that we first
select the appropriate database strategy, and
only then act on it. This also means we will
always have the same metrics even if load-
balancing is disabled.