repository: Convert #fetch_upstream to use in-memory remotes
Convert `Repository#fetch_upstream()` to use `fetch_remote()` with an
in-memory remote. This allows us to skip the call to `add_remote()`: its
backing RPC call is deprecated and will eventually be removed by Gitaly.

The change reuses the `fetch_remote_params` feature flag which is also
used by the GitHub importer and by `fetch_as_mirror()`. This is fine
given that the feature flag hasn't yet been rolled out to production and
is thus still disabled.