Reset pagination on sort change on Releases page
This commit fixes an issue with the Releases page that was caused by
switching the sort direction after paginating the page.

This fix avoids this issue by always reseting the pagination cursors
("before", "after") when the sort is changed, effectively placing the
user back on page 1 of the results.

Changelog: fixed