Force ASCII-8BIT encoding in CI job trace
This commit adds a feature flag, ci_job_trace_force_encode, that will
ensure that CI job traces can be appended when a UTF-8 string arrives by
forcing the encoding.

We also track the exception with the data to understand the source of
the problem.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/333424

Changelog: fixed