Use database as SSOT for diffs whitespace
This MR removes most of the code for using cookies
and querystrings to set whether or not to show whitespace
diffs in merge requests.

Then saves the changes to the database directly from the
settings dropdown on the diffs page. This keeps the user preferences
in sync and makes them the single source of truth for displaying
whitespace.

Changelog: changed