Implement in-memory remotes for `update_remote_mirror`
The `update_remote_mirror` functions is currently using on-disk remotes
by first calling `remote_mirror.ensure_remote!` followed by the call to
`update_repository`. Gitaly is phasing out support for on-disk remotes
though in favor of in-memory remotes, where the remote is specified as
part of the request itself.

Implement the ability to use in-memory remotes for this RPC. For now,
this is implemented behind a feature flag.