Fix new users receiving a failed pipeline warning without CI YAML
When new users create a project without a CI YAML, they previously would
receive a warning about needing credit card validation. This happened
because the credit card validation check would happen earlier in the
sequence than the YAML content processing.

To fix this, we add a new sequence in the chain that checks the credit
card only after a valid pipeline will be created. We keep the
`Abilities` check where it is because the existence of a deleted project
and other cheaper checks should happen early in the sequence.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/331959

Changelog: fixed