Move webpack_public_path logic to a rails helper
We are currently setting a global variable `__webpack_public_path__` in
a global JavaScript function that gets the value from the `Gon` object.
This just happens in the page specific JavaScript. The value from this
variable is used for e.g. dynamically loading JavaScript assets like the
WebIDE or parts of the PerformanceBar.

The problem is that the PerformanceBar is currently loaded _before_ this
bit of JavaScript is executed so webpack doesn't have the correct path
set. While investigating this issue, it also appears that it could
affect other JavaScript bundles (e.g. Sentry) or if a page uses a
different layout (e.g. the Jira Connect App iframe.

In order to fix this globally we are moving the logic to be executed
directly _after_ an include of the global `Gon` object. In order to
ensure that one cannot be used without the other, we are moving the
initialization of `Gon` _and_ the `__webpack_public_path__` to a new
helper called: `render_gon_and_global_window_variables`.

This new helper `render_gon_and_global_window_variables` could be
utilized for other global JavaScript as well (e.g. window.gl).