Fix Sidekiq mailer specs
These specs were using a code snippet copied from a GitHub issue
comment, but I missed our resetting the test adapter correctly. Also
move this to spec metadata so we can easily reuse it elsewhere if we
need to.

The reason we need this anyway is because we can't easily set a non-test
adapter for ActiveJob in tests: either for a specific test, or
globally. (Request specs will get set back to the test adapter.)