Update administration pages with new admin area access info
GitLab 14.0 includes an interface change that moves the UI
element to access the Admin Area to a new location;
from the top menu bar to within the Menu option on the top
menu bar.

This updates the administration documentation pages that describe
accessing the Admin Area to reflect the new interface change.