Fix double counting Sidekiq queue events
In those circumstances where a Sidekiq job is enqueued
in the future e.g. via perform_in, the Sidekiq counting
job enqueue events was being called twice: first for
the initial call to perform_in, and again when Sidekiq
actually queued it for execution after the elapsed delay.

We now detect this and skip jobs in this middleware that
have the `at` field set.

Changelog: fixed