Remove check for gitlab employee badge feature
The `#gitlab_employee?` method would return false if the
Feature flag `gitlab_employee_badge` was enabled. Since this feature
flag only added a special badge to gitlab employees and didn't affect
the actual check, we're removing it. It was actually harming spam checks
which will allow issue creation without checking for spam if the user is
a gitlab employee. Since this feature flag is disabled, it always
failed.