Perform CI trace cleanup in further background jobs
The archive trace cleanup cron (that catches otherwise interrupted/broken CI traces archiving) runs once an hour.  However large traces can take > 10 minutes to collate into an artifact (reading 10K+ chunks from object storage at ~15 chunks/s) and this job can get interrupted by kubernetes operations.  Once a large job is at the head of the queue for this cleanup worker, it may take many attempts for this to complete (or it might *never* be able to complete).

By spinning out each individual cleanup into a separate background job (the same job that tried to run when the job finished in the first instance):
1. All the jobs with sufficiently small traces will get a good chance to be archived, clearing the field (and chunks table in the DB) somewhat
1. With retries on the individual trace archive there's perhaps a chance that more will get processed (vs the cron worker being retried and running into locks on the builds).

This is only a first mitigation to reduce the number of affected builds; because the problem builds are large (by definition), they will have the bulk of the impact on the size of ci_build_trace_chunks, and that will need separate mitigation (trying to make the archive job faster or re-startable)

See https://gitlab.com/gitlab-org/gitlab/-/issues/330141#note_603412446 for some related analysis of the impact this is having currently