gitaly_client: Allow BlobsStitcher to process blobs with no mode or path
The BlobsStitcher object will reassemble blobs which have been split up
across multiple gRPC messages. It is only used in the context of the
`get_blobs()` RPC right now, which has a bunch of information about
blobs which may not be available via other RPCs. While the stitcher
gracefully handles absence for many of the parameters, absence of a blob
mode and path will cause failure.

Relaxe the stitcher to work alright in case either mode or path are
absent. This is in preparation for the introduction of the
`list_blobs()` RPC.