Update profile information summary for on-demand scans
This commit solves: https://gitlab.com/gitlab-org/gitlab/-/issues/290081
It removes the ScanType from the Scanner Profile selector.
It removes the ValidationStatus from site Profile Selector.
Adds the Validation Status to the site Profile section.
Adds "None" if optional fields have no value.
All that is to be found on the on-demand scans page.

Changelog: changed