Track usage data for Network Policies UI
Uses a redis HLL counter to track the number of clusters visiting the
network policies page with network policies present in the cluster.
Since any ultimate customer can visit the page, we require policies to
be present in the response as an extra signal for the feature having
provided value to the user.

See https://gitlab.com/gitlab-org/gitlab/-/issues/334090

Changelog: added