Save last_synced_at on licenses table
Add a column for last_synced_at to the licenses table.
Use this new column for the "Last sync" value shown in the details
section of the Subscription page.
Modify the activate and sync services to save last_synced_at when
creating a cloud license record.

Changelog: added
EE: true