Add clarifying docs about sidekiq queue migration
https://gitlab.com/gitlab-com/gl-infra/production/-/issues/4957 happened
because such a migration happened in the normal migration, not a
post-deploy migration.  There are docs for this (in the post-deployment
migration page), but clearly that was (and remains) easy to miss.
Calling it out specifically and clearly in the location where someone
will be finding the need for the migration seems useful.