Implement caching on discussions.json for merge requests
This is improve performance of discussions.json by caching.

Locally, when request hits a warm cache, it shows improvement
from 10s to 1.5s for 100+ discussions. For 100+ discussions, each
request reads `203569` bytes.

This is implemented behind `merge_request_discussion_cache` feature
flag and only works for merge requests. This is to see how it is
going to perform on production and to ensure we can turn the feature
off if there are unexpected bugs.