Add metrics on searching for caught up replicas
Add new metrics to reflect how often to we search for caught up replicas
(in the current implementation, we demand that all of the replicas
should be caught up to unstick). It would allow us to understand how
often do we perform that operation and the distribution of the results.

Changelog: added