Generate diff rows array on load instead of in computed property
Seeing as most of this data doesn't change
we can generate this on load and save it to the state
this allows us to freeze a lot of the data later on to help with memory.
But what this also means is that we don't have a computed
property update any time a diff line updates, we can just
update that single object which then creates 1 update, instead of
X updates (X is number of diff lines in file).

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/333691