Fix admin mode when authenticating with LDAP
When users reauthenticate to enable admin mode, and are signing
in with an external authentication provider such as LDAP, we need
to use a case-insensitive, normalized lookup of the user identity.
This changes the match identity method in the user model to use
the existing scope which normalizes the uid and does a case-insensitive
lookup.

Changelog: changed