Fix toggling of shared runners when credit card is needed
This fixes a minor UI bug where when after validating an account,
toggling the shared runners off again will cause the validation error to
display again. Users can workaround this by reloading the page.

To avoid this problem, once we've successfully validated, we store that
state and use it to determine to show the warning. This also simplifies
the code.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/334322