Delete failed releases before deploying
When a deployment fails GitLab retries the review-deploy job, but if
the failure is fatal it will merely fail three times before giving up.

This way review-deploy will check if a failed deployment is present,
delete it, and then try again from scratch.

The disadvantage is that if a deployment fails for a transient reason
it is less efficient to delete before deploying again.