Refactor Gitlab::Database to support multiple DBs
This refactors the module Gitlab::Database so it supports multiple
databases. Logic tied to a connection is now scoped to instances of
Gitlab::Dababase::Database. For every database, an instance of this
class is created and stored in Gitlab::Database::DATABASES.

See https://gitlab.com/gitlab-org/gitlab/-/issues/331776 for more
information.