Only render FileFinderOverlay when needed
We currently always render the FileFinderOverlay, even when a user
doesn't use it. Depending on the size of a Merge Request / Repo, this
can safe hundreds of DOM-Nodes from being rendered unnecessarily.

For example on a simple Web IDE page around 450 of 2200 DOM nodes are
inside this modal [0]. On a large Merge Request (241 changed files) [1]
around 800 DOM nodes are attributed to this modal.

A simple way to test how many unnecessary DOM nodes are rendered:

```js
 document.getElementsByTagName('*').length
 document.querySelector('.file-finder-overlay').remove()
 document.getElementsByTagName('*').length
```

[0]: https://gitlab.com/-/ide/project/gitlab-com/www-gitlab-com/edit/master/-/
[1]: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/50464/diffs

Changelog: performance