Add error column to background_migration_jobs
Add a new column to background_migration_jobs where an error message can
be stored, and add a new model method to simplify populating the value.
For now, this requires manual usage, but in the future an hook can be
added in BG jobs that will automatically save exception information.

Changelog: changed