Refactor: Split UpdateBuildsMinute Service
We want to cancel pipelines before deleting them because minutes are not accumulated on deleted jobs that continue to run.  In order to cancel pipelines before deleting them and have minutes accumulate properly, we need to compute the minutes synchronously before the destroy to pass of to an async worker that can do the rest.

This change breaks apart the UpdateBuildsMinute service into 2 so that part of the logic can be executed async. There are no changes to the existing behavior in this code.