Skip the QA tunnel when using k3s in tests
The :kubernetes tag configures the GitLab instance under test to use
the QA tunnel for external access. This isn't necessary when using k3s
for the kubernetes cluster.

Then we don't need the :orchestrated tag either because it works with
the default GitLab configuration and only additional docker containers
that the test framework manages (not gitlab-qa).

We ensure the cluster is removed in the after hook, just in case there
are more errors when removing the runner.