Refactor license validation specs
Prior to this change there were several different styles for declaring
the top level describe block in the License validation specs. This made
it difficult to easily see which validation is being tested where.

This commit re-groups the existing validation specs in this file to be
under a consistent title (the validation method name).

Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/334115