Remove Redis trace chunks feature flag
This flag controlled which Redis was used for trace chunks - the
persistent Redis or the trace chunks Redis. For every instance except
GitLab.com, these are the same instance. We are deliberately not
documenting the trace chunks instance[1] as this adds operational
complexity for very little gain unless you are at the size of
GitLab.com.

We can still document this instance in future, if we choose to.

[1]: https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/1131

Changelog: other