Add docs for backing up git data with direct disk access
The reason behind this is backing up projects in fork networks,
as the regular backup variants will duplicate the already-deduplicated
objects in pool repositories for each backed up project in the fork
network.

This describes the issue and introduces an experimental workaround,
by copying the repositories as-is on disk, while preventing writes.