Reset the current page to 1 on tab / sort change
When opening a different tab (for example, from Open
to Closed), we should reset the page to 1 to avoid
showing "No results found" page when there are less
issues in a particular tab.

The same applies to sort since it doesn't make sense
to sort a list but leave you on page 2 (which might be in
the middle of the list).