Migrate New Environments Form to Vue
To allow for richer frontend validations and to align better with the
pajamas design system, I've migrated the new environments form to vue.

I've decided to do a full migration rather than using
`parseRailsFormFields` because I found out about it too late, although
it would have likely been easier.

The form migration is about as bare-bones as possible, no vuex or other
state management, and I use the existing internal API for speed.

Next is to port the Edit Environment form, which is happening separately
as it might involve a little bit of vuex or other state management. As
it is only 2 fields, I suspect it won't, but we will see.

Changelog: changed